Добавени функции open и saveas

+++ src/files/WriteFile.java
@@ -2,16 +2,38 @@ package files;
 
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
 
 public class WriteFile extends FileCreate {
     private FileWriter fileWriter;
-    public WriteFile(String fileName) throws IOException {
-        super(fileName);
-        super.createFile();
-        fileWriter=new FileWriter(file);
+    public Path pathName;
+
+    public WriteFile(String fileName, String path) throws IOException {
+        super(fileName, path);
+        checkFile();
+        pathName= Paths.get(path);
+        if(Files.exists(pathName)) {
+            fileWriter = new FileWriter(file);
+        }
+        else {
+            System.out.println("Тази пътека не съществува");
+        }
     }
 
-    public void writeInFile(String content) throws IOException {
+    @Override
+    public void checkFile() throws IOException {
+        if(file.createNewFile()){
+            System.out.println("Файлът е създаден успешно");
+        }
+        else {
+            System.out.println("Файлът вече съществува. Данните са заместени успешно");
+        }
+    }
+
+    public void action(String content) throws IOException {
         fileWriter.write(content);
         fileWriter.close();
     }
